pyproject.toml
[project]
name = "video-downloader"
version = "1.0.0"
description = "多平台视频下载和文案提取工具 - 支持抖音、快手、小红书、B站、YouTube等"
readme = "README.md"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "aiohttp>=3.9.0",
    "aiofiles>=23.2.0",
    "yt-dlp>=2023.12.30",
    "openai>=1.6.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.2",
    "you-get>=0.4.1743",
    "tqdm>=4.65.0",
    "lxml>=4.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
video-dl = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dependencies = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0"]

# 由于项目结构是扁平的，需要配置hatch来正确处理
[tool.hatch.build.targets.wheel]
packages = ["."]

# 开发工具配置
[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
target-version = "py312"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]